Support Dinghy wildcard domains
Dinghys wildcard syntax is just prefix with a dot (`.example.com`) where
as mkcert uses an asterisk (`*.example.com`).

So we rewrite the `VIRTUAL_HOST` if it starts with a dot.

To do that we eed to split up the compound onliner finding the domains
in either `MKCERT_DOMAINS`, `VIRTUAL_HOST` or hostname. But that is
probably better for reading and understanding anyway (at least it is
easier to document all steps now).