Enable the mod_headers module
This change makes it possible to add CORS and other headers to HTTP responses.

This is necessary if we want to use static assets like fonts and icons
in single directory components rendered in Storybook using the new
Storybook contrib module.

Add the following directives to the `.htaccess`-file in your web root to
set the required header:

```
<FilesMatch "\.(eot|gif|ico|jpeg|jpg|otf|png|svg|ttf|woff|woff2)$">
  <IfModule mod_headers.c>
    Header set Access-Control-Allow-Origin "*"
  </IfModule>
</FilesMatch>
```